Move struct to a new folder
- Update code comments.
